// src/app/lib/types.ts
// types.ts
export interface Developer {
  id: number;
  full_name: string;
  age: number;
  birth_date: string;
  phone_number: string;
  nacionality: string;
  summary: string;
  stack: string;
  main_stack_technology: string;
  profile_image: string;
  linkedin_profile: string;
  github_profile: string;
}